Implemented plugins approach for SiteResolver
