+++ packages/create-sitecore-jss/src/templates/nextjs-multisite/src/lib/site-resolver/plugins/multisite.ts
@@ -1,0 +1,14 @@
+import { SiteInfo } from '@sitecore-jss/sitecore-jss-nextjs/middleware';
+import config from 'temp/config';
+import { SiteResolverPlugin } from '..';
+
+class MultisitePlugin implements SiteResolverPlugin {
+  exec(sites: SiteInfo[]): SiteInfo[] {
+    // Add preloaded sites
+    sites.push(...(JSON.parse(config.sites) as SiteInfo[]));
+
+    return sites;
+  }
+}
+
+export const multisitePlugin = new MultisitePlugin();
