+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/site-resolver/plugins/default.ts
@@ -1,0 +1,18 @@
+import { SiteInfo } from '@sitecore-jss/sitecore-jss-nextjs/middleware';
+import config from 'temp/config';
+import { SiteResolverPlugin } from '..';
+
+class DefaultPlugin implements SiteResolverPlugin {
+  exec(sites: SiteInfo[]): SiteInfo[] {
+    // Add our single site
+    sites.unshift({
+      name: config.jssAppName,
+      language: config.defaultLanguage,
+      hostName: '*',
+    });
+
+    return sites;
+  }
+}
+
+export const defaultPlugin = new DefaultPlugin();
